[Bugfix] Make sure we get the failing arguments on a timeout
There was a weird issue which arises from a strange design decision in
QuickCheck, which means that we don't get the arguments that cause the
failure if the check times out and the within is not within the body of
the property. So now we inject the `within` right before the body.