[Feature] Genetic Program Repair
By running the `repairAttempt` in a loop and "breeding" the good
solutions with each other, we can fix programs with multiple bugs, such
as those seen in TwoFixes, ThreeFixes etc... so now we can finally use
the name HenProg!

This commit also creates a new test-suite, slow-tests, so that we can
test the new genetic repair separately.