[Feature] Allow use of evalutaed leaves only for fault-localization
This can be very good in cases where the bug is in a leaf, e.g.
FourFixes or other code where it is nested in some big expression
and we only want to change that one thing. It does however limit
our flexibility a lot: e.g. BrokenGCD will now fail to repair
since the neccesary fix is not to change a leaf!